data-collection/requirements.txt
# FastAPI and API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for API calls
aiohttp==3.9.1
httpx==0.25.2

# Async file operations
aiofiles==23.2.0

# Web scraping - Core dependencies
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
webdriver-manager==4.0.1

# Web scraping - Additional parsing libraries
html5lib==1.1
cssselect==1.2.0

# Environment and configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging
loguru==0.7.2

# Date/time handling
python-dateutil==2.8.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0

# Production server (optional)
gunicorn==21.2.0

# Core system dependencies (auto-installed by other packages but listed for clarity)
urllib3==2.1.0
certifi==2023.11.17
click==8.1.7
typing-extensions==4.8.0 